refactoring - part 3, delete sequential code
