+++ polarityjam/polarityjam/model/masks.py
@@ -1,14 +1,13 @@
 import numpy as np
 import skimage.filters
-from cellpose import utils
 from scipy import ndimage as ndi
 
 from polarityjam.polarityjam_logging import get_logger
 from polarityjam.utils import parameters
-from polarityjam.utils.compute import otsu_thresh_mask
+from polarityjam.compute.compute import otsu_thresh_mask
 
 
-class SingleCellMasks:
+class SingleCellMasksCollection:
 
     def __init__(self, masks, connected_component_label):
         self.connected_component_label = connected_component_label
@@ -84,7 +83,7 @@ class SingleCellMasks:
         return outline_mask
 
 
-class Masks:
+class MasksCollection:
     def __init__(self, cell_mask):
         self.cell_mask = cell_mask
         self.cell_mask_rem_island = None
@@ -140,69 +139,9 @@ class Masks:
         return organelle_mask
 
 
-def get_single_cell_membrane_mask(parameters, im_marker, im_junctions, single_cell_mask):
-    """Gets the single cell membrane mask."""
-    if im_marker is not None or im_junctions is not None:
-        return get_outline_from_mask(single_cell_mask.astype(bool), parameters["membrane_thickness"])
-    return None
-
-
 def get_single_cell_mask(connected_component_label, cell_mask):
     """Gets the single cell mask from a mask where each cell has an increasing connected component value."""
     return np.where(cell_mask == connected_component_label, 1, 0)  # convert connected_component_label to 1
-
-
-def get_single_cell_organelle_mask(connected_component_label, organelle_mask):
-    """Gets the single cell organelle mask."""
-    if organelle_mask is not None:
-        return np.where(organelle_mask == connected_component_label, 1, 0)
-    return None
-
-
-def get_single_cell_cytosol_mask(single_cell_mask, im_marker, single_nucleus_mask):
-    """Gets the cytosol mask."""
-    if single_nucleus_mask is not None and im_marker is not None:
-        return np.logical_xor(single_cell_mask.astype(bool), single_nucleus_mask.astype(bool))
-    return None
-
-
-def get_single_cell_nucleus_mask(connected_component_label, nuclei_mask):
-    """Gets the single cell nucleus mask."""
-    if nuclei_mask is not None:
-        return np.where(nuclei_mask == connected_component_label, 1, 0)
-    return None
-
-
-def get_organelle_mask(parameters, img, cellpose_mask):
-    """Gets the organelle mask."""
-    if parameters["channel_organelle"] >= 0:
-        img_organelle_blur = ndi.gaussian_filter(img[:, :, parameters["channel_organelle"]], sigma=3)
-
-        organelle_mask = np.where(img_organelle_blur > skimage.filters.threshold_otsu(img_organelle_blur), True, False)
-        organelle_label = organelle_mask * cellpose_mask
-
-        return organelle_label
-    return None
-
-
-def get_single_junction_protein_mask(single_membrane_mask, im_junction):
-    if im_junction is not None:
-        single_cell_junction_protein = otsu_thresh_mask(single_membrane_mask, im_junction)
-        return single_cell_junction_protein.astype(bool), single_cell_junction_protein
-    return None, None
-
-
-def get_nuclei_mask(parameters, img, cellpose_mask):
-    """Gets the nuclei mask."""
-    if parameters["channel_nucleus"] >= 0:
-        img_nuclei_blur = ndi.gaussian_filter(img[:, :, parameters["channel_nucleus"]], sigma=3)
-
-        nuclei_mask = np.where(img_nuclei_blur > skimage.filters.threshold_otsu(img_nuclei_blur), True, False)
-
-        nuclei_label = nuclei_mask * cellpose_mask
-
-        return nuclei_label
-    return None
 
 
 def get_outline_from_mask(mask, width=1):
@@ -214,17 +153,3 @@ def get_outline_from_mask(mask, width=1):
     outline_mask = np.logical_xor(dilated_mask, eroded_mask)
 
     return outline_mask
-
-
-def get_outline_with_multiple_labels(mask, width=1):
-    """ TODO: not used at the moment, faster for multiple masks/mask labels"""
-
-    outline_list = np.array(utils.outlines_list(mask))
-    outlines = np.zeros((mask.shape[0], mask.shape[1]))
-    for mask_id, outline_coords in enumerate(outline_list):
-        if outline_coords.T.shape[1] < 10:
-            outlines[tuple(outline_coords.T)] = mask_id + 1
-
-    outlines_mask = ndi.morphology.binary_dilation(outlines.astype(bool), iterations=width)
-
-    return outlines_mask
